backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.2.0
bcrypt>=4.1.0
PyJWT>=2.9.0
Pillow>=10.0.0
